requirements-backend.txt
# requirements-backend.txt
# Slim dependencies for the FastAPI backend (Docker / HF Spaces).
# For the full notebook/experiment stack, use requirements.txt instead.

# FastAPI + server
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.9
aiofiles>=23.0.0

# Database
sqlalchemy>=2.0.0

# Validation / config
pydantic
python-dotenv==1.2.1

# Document parsing (policy analyzer)
pypdf==6.7.0
python-docx>=1.1.0

# LangChain / LangGraph (policy analyzer pipeline)
langchain-core
langchain-openai==1.1.7
langgraph==1.0.5

# Observability / Tracing (Arize)
opentelemetry-sdk>=1.39.0
opentelemetry-exporter-otlp-proto-grpc>=1.39.0
openinference-instrumentation-langchain>=0.1.58
grpcio>=1.60.0

# HTTP client (tests + general)
httpx>=0.27.0